factor out ClampedHTML to demo control
