+++ stories/explorer/basic.js
@@ -11,7 +11,7 @@ import { getESSchemas } from '../../src/utils/schema'
 import ExampleTypes from '../../src/exampleTypes/'
 import { FilterList } from '../../src/FilterList'
 import { Adder, Pager } from '../DemoComponents'
-import { Input } from '../DemoControls'
+import { Input, ClampedHTML } from '../DemoControls'
 let { ResultCount, ResultTable, TypeMap } = ExampleTypes({ Input })
 
 import Contexture, { es, schemas, updateClient } from './contexture'
@@ -56,17 +56,6 @@ let changeSchema = schema => {
     ],
   })
 }
-
-let textTruncate = {
-  overflow: 'hidden',
-  textOverflow: 'ellipsis',
-  // WebkitLineClamp: '4',
-  // WebkitBoxOrient: 'vertical',
-  maxHeight: '100px',
-}
-let ClampedHTML = x => (
-  <div style={textTruncate} dangerouslySetInnerHTML={{ __html: x }} />
-)
 
 let lookups = {
   display: {
